Reworking of the code per Mark's comments
Mark didn't care for the type punning on unit8_t *; he rightly said that in
some cases, a compiler can align such an array on the stack at any arbitrary
location, so this *could* be bad on architectures where alignment is
strictly enforced.

To address this issue, I created a type dns_align_t to declare buffers with
proper alignment (if you go the stack route).  dns.h has some sample
declarations with this type.

I really don't want to lose this method as I feel it keeps things coherent
in the cache.