Fixed EDNS0 support.
I really screwed the pooch when reading RFC-2671.  It felt that RFC-2671 was
a bit ambiguous and yup, I mis-read it.  Basically, the extended lables did
not start an OPT RR, but was a mechanism to expand on the number of label
types in DNS.  It still requires some special processing, because the
standard RR fields have all new meanings, but at least it isn't as nasty as
I originally thought.