Remove unnecessary pointer modifiers.
I've rethought my stance on const pointers.  It doesn't buy you much, and
the semantics of it don't (I think) withstand scrutany.  Also, the restrict
keyword is only required for pointer parameters of the same type, so they
mostly go.

I'll be doing this mostly in a later code reorganization, but for now, let's
start here.