Bug fix---add support for Windows/mingw
Thanks to Alper Akcan for reporting this issue and sending the patch.  I
didn't accept the patch as is as it included much white space patches I
didn't want, and what I did want didn't match the coding style I use.  But
the patch that was required was small enough to add by hand.

There was also more changes required for a 2.0.1 release, like updating the
Makefile and Lua rockspec.